requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
openai>=1.30.0
groq>=0.8.0
pymilvus>=2.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
sentence-transformers>=2.7.0
asyncio-mqtt>=0.16.0
sqlalchemy>=2.0.0
pgvector>=0.2.5
tenacity>=8.3.0
loguru>=0.7.2
